+++ src/App/DavLockManager.py
@@ -12,18 +12,13 @@
 ##############################################################################
 
 from AccessControl.class_init import InitializeClass
+from AccessControl.Permissions import webdav_manage_locks
 from AccessControl.SecurityInfo import ClassSecurityInfo
 from Acquisition import Implicit
 from Acquisition import aq_base
 from App.special_dtml import DTMLFile
 from OFS.Lockable import wl_isLocked
 from OFS.SimpleItem import Item
-
-
-try:
-    from webdav import webdav_manage_locks
-except ImportError:  # ZServer shadows our own webdav module
-    from OFS.bbb import webdav_manage_locks
 
 
 class DavLockManager(Item, Implicit):